requirements.txt
gradio
neo4j
nltk
numpy
openai
openpyxl
pandas
python-dotenv
requests
tqdm